Encapsulated trip session and replay trip session management in the TripSessionComponent. Moved ReplayRouteTripSession to dropin module.
