+++ libnavui-app/src/test/java/com/mapbox/navigation/ui/app/internal/controller/TripSessionStarterStateControllerTest.kt
@@ -1,24 +1,12 @@
 package com.mapbox.navigation.ui.app.internal.controller
 
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
-import com.mapbox.navigation.core.MapboxNavigation
-import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.testing.MainCoroutineRule
-import com.mapbox.navigation.ui.app.internal.State
-import com.mapbox.navigation.ui.app.internal.navigation.NavigationState
 import com.mapbox.navigation.ui.app.internal.tripsession.TripSessionStarterAction
-import com.mapbox.navigation.ui.app.internal.tripsession.TripSessionStarterState
 import com.mapbox.navigation.ui.app.testing.TestStore
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.mockkObject
 import io.mockk.spyk
-import io.mockk.unmockkAll
-import io.mockk.verify
-import io.mockk.verifyOrder
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.runBlockingTest
-import org.junit.After
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -30,163 +18,32 @@ class TripSessionStarterStateControllerTest {
     var coroutineRule = MainCoroutineRule()
 
     private lateinit var testStore: TestStore
+    private lateinit var sut: TripSessionStarterStateController
 
     @Before
     fun setup() {
-        mockkObject(MapboxNavigationApp)
         testStore = spyk(TestStore())
-    }
-
-    @After
-    fun teardown() {
-        unmockkAll()
+        sut = TripSessionStarterStateController(testStore)
     }
 
     @Test
-    fun `startTripSession if location permissions are granted`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    tripSession = TripSessionStarterState(
-                        isLocationPermissionGranted = false,
-                        isReplayEnabled = false,
-                    )
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-            val mapboxNavigation = mockMapboxNavigation()
+    fun `on OnLocationPermission action should update TripSessionStarterState`() {
+        testStore.dispatch(TripSessionStarterAction.OnLocationPermission(true))
 
-            sut.onAttached(mapboxNavigation)
-            testStore.dispatch(
-                TripSessionStarterAction.OnLocationPermission(true)
-            )
-
-            verify { mapboxNavigation.startTripSession() }
-        }
-
-    @Test
-    fun `onDetached does not stopTripSession for a regular session`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    tripSession = TripSessionStarterState(
-                        isLocationPermissionGranted = true,
-                        isReplayEnabled = false,
-                    )
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-            val mapboxNavigation = mockMapboxNavigation()
-
-            sut.onAttached(mapboxNavigation)
-            sut.onDetached(mapboxNavigation)
-
-            verify(exactly = 1) { mapboxNavigation.startTripSession() }
-            verify(exactly = 0) { mapboxNavigation.stopTripSession() }
-        }
-
-    @Test
-    fun `EnableTripSession will restart a trip session when replay is enabled`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    navigation = NavigationState.ActiveNavigation,
-                    tripSession = TripSessionStarterState(
-                        isLocationPermissionGranted = true,
-                        isReplayEnabled = true,
-                    )
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-            val mapboxNavigation = mockMapboxNavigation()
-
-            sut.onAttached(mapboxNavigation)
-            testStore.dispatch(TripSessionStarterAction.EnableTripSession)
-
-            verifyOrder {
-                mapboxNavigation.startReplayTripSession()
-                mapboxNavigation.stopTripSession()
-                mapboxNavigation.startTripSession()
-            }
-        }
-
-    @Test
-    fun `EnableReplayTripSession will startReplayTripSession`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    navigation = NavigationState.ActiveNavigation
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-            val mapboxNavigation = mockMapboxNavigation()
-
-            sut.onAttached(mapboxNavigation)
-            testStore.dispatch(
-                TripSessionStarterAction.OnLocationPermission(true)
-            )
-            testStore.dispatch(TripSessionStarterAction.EnableReplayTripSession)
-
-            verify { mapboxNavigation.startReplayTripSession() }
-        }
-
-    @Test
-    fun `EnableReplayTripSession will not startReplayTripSession without location permissions`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    navigation = NavigationState.ActiveNavigation,
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-            val mapboxNavigation = mockMapboxNavigation()
-
-            sut.onAttached(mapboxNavigation)
-            testStore.dispatch(
-                TripSessionStarterAction.OnLocationPermission(false)
-            )
-            testStore.dispatch(TripSessionStarterAction.EnableReplayTripSession)
-
-            verify(exactly = 0) { mapboxNavigation.startReplayTripSession() }
-        }
-
-    @Test
-    fun `EnableReplayTripSession will only startReplayTripSession for ActiveGuidance`() =
-        runBlockingTest {
-            testStore.setState(
-                State(
-                    navigation = NavigationState.FreeDrive,
-                    tripSession = TripSessionStarterState(
-                        isLocationPermissionGranted = true,
-                        isReplayEnabled = true,
-                    )
-                )
-            )
-            val sut = TripSessionStarterStateController(testStore)
-
-            val mapboxNavigation = mockMapboxNavigation()
-
-            sut.onAttached(mapboxNavigation)
-            testStore.setNavigationState(NavigationState.DestinationPreview)
-            testStore.setNavigationState(NavigationState.RoutePreview)
-            testStore.setNavigationState(NavigationState.Arrival)
-
-            verify(exactly = 0) { mapboxNavigation.startReplayTripSession() }
-            testStore.setNavigationState(NavigationState.ActiveNavigation)
-            verify(exactly = 1) { mapboxNavigation.startReplayTripSession() }
-        }
-
-    private fun mockMapboxNavigation(): MapboxNavigation {
-        val mapboxNavigation = mockk<MapboxNavigation>(relaxed = true)
-        every { MapboxNavigationApp.current() } returns mapboxNavigation
-        return mapboxNavigation
+        assertEquals(true, testStore.state.value.tripSession.isLocationPermissionGranted)
     }
 
-    private fun TestStore.setNavigationState(navState: NavigationState) {
-        setState(
-            state.value.copy(
-                navigation = navState
-            )
-        )
+    @Test
+    fun `on EnableReplayTripSession action should update TripSessionStarterState`() {
+        testStore.dispatch(TripSessionStarterAction.EnableReplayTripSession)
+
+        assertEquals(true, testStore.state.value.tripSession.isReplayEnabled)
+    }
+
+    @Test
+    fun `on EnableTripSession action should update TripSessionStarterState`() {
+        testStore.dispatch(TripSessionStarterAction.EnableTripSession)
+
+        assertEquals(false, testStore.state.value.tripSession.isReplayEnabled)
     }
 }
