+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/util/TestStore.kt
@@ -10,4 +10,8 @@ internal class TestStore : Store() {
     fun setState(state: State) {
         _state.value = state
     }
+
+    fun updateState(update: (State) -> State) {
+        setState(update(_state.value))
+    }
 }
